+++ RyuSocks/Packets/Auth/Constants.cs
@@ -18,6 +18,6 @@ namespace RyuSocks.Packets.Auth
 {
     public static class Constants
     {
-        public const byte UaPVersion = 0x01;
+        public const byte UsernameAndPasswordVersion = 0x01;
     }
 }